Move Inkuire from ClassLikeSupport
